refactor(Cli): Regroup all cli files in job_scrapper/cli #51
