--- job_scrapper/cli/cli_configure.py
@@ -1,5 +1,0 @@
-import cloup
-
-@cloup.group()
-def configure():
-    pass
