+++ job_scrapper/cli/configure/cli_localisation.py
@@ -3,7 +3,7 @@ import cloup
 from job_scrapper.scrapper_skeleton.scrapper_skeleton import JobScrapperSkeleton
 from job_scrapper.sql.tables import Places, Distances
 from sqlalchemy import or_
-from job_scrapper.cli.ask_contact import CONTACT_OPTION, ask_contact
+from job_scrapper.cli.configure.cli_contact import CONTACT_OPTION, ask_contact
 from job_scrapper.tools.geolocalisation import Geolocalisation
