+++ job_scrapper/cli/standalones/request_offer_job_cli.py
@@ -1,6 +1,6 @@
 import click
 import cloup
-from job_scrapper.cli.offers_cli import (
+from job_scrapper.cli.offers.offers_cli import (
     COLUMN_OPT,
     DISTANCE_OPT,
     KEYWORD_OPT,
